fix: change when monorepo tests
